Refactor: change dtos to dto
