+++ apps/api-server/src/modules/auth/auth.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get, Header, HttpCode, Post, Redirect, Req, UseGuards } fro
 import { AuthGuard } from '@nestjs/passport';
 import { User } from '@src/entities/users.entity';
 import { AuthService } from './auth.service';
-import { UserKakaoDto } from './dtos/users.kakao.dto';
+import { UserKakaoDto } from './dto/users.kakao.dto';
 import { JwtRefreshGuard } from './guards/jwt-refresh-auth.guard';
 
 @Controller('auth')
