+++ apps/api-server/src/modules/auth/strategies/kakao.strategy.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common';
 import { OauthConfig } from '@src/config/config.constant';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-kakao';
-import { UserKakaoDto } from '../dtos/users.kakao.dto';
+import { UserKakaoDto } from '../dto/users.kakao.dto';
 
 @Injectable()
 export class KakaoStrategy extends PassportStrategy(Strategy) {
